pyproject.toml
[project]
name = "dedup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "opencv-python>=4.12.0.88",
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
]
